Move preparations into a separate file
