+++ Sources/SteamPress/Controllers/BlogAdminController.swift
@@ -5,6 +5,7 @@ import AuthProvider
 import Foundation
 import Fluent
 import Validation
+import Cookies
 
 struct BlogAdminController {
 
@@ -378,25 +379,33 @@ struct BlogAdminController {
         }
         
         let passwordCredentials = Password(username: username.lowercased(), password: password)
+//        let rememberMeCookie = Cookie(name: "steampress-remember-me", value: "")
         
-        if rememberMe {
-            request.storage["remember_me"] = true
-        }
-        else {
-            request.storage.removeValue(forKey: "remember_me")
-        }
+//        if rememberMe {
+//            
+//            request.cookies.insert(rememberMeCookie)
+//            request.storage["remember_me"] = true
+//        }
+//        else {
+//            request.storage.removeValue(forKey: "remember_me")
+//        }
         
         do {
             let user = try BlogUser.authenticate(passwordCredentials)
             request.auth.authenticate(user)
-            request.storage.removeValue(forKey: "remember_me")
+//            request.storage.removeValue(forKey: "remember_me")
 
-            return Response(redirect: pathCreator.createPath(for: "admin"))
+            let response = Response(redirect: pathCreator.createPath(for: "admin"))
+            if rememberMe {
+                response.storage["remember_me"] = true
+            }
+            return response
         }
         catch {
             log.debug("Got error logging in \(error)")
             let loginError = ["Your username or password was incorrect"]
-            request.storage.removeValue(forKey: "remember_me")
+//            request.storage.removeValue(forKey: "remember_me")
+//            request.cookies.remove(rememberMeCookie)
             return try viewFactory.createLoginView(loginWarning: false, errors: loginError, username: username, password: "")
         }
     }
